#2: Use ConfigOptions instead of defining --config param everywhere.
